refactor: extract fake logs creation methods
